refactor: session request response 패키지 분리
