+++ tomcat/src/main/java/nextstep/jwp/handler/HttpRequestHandler.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.handler;
 
-import nextstep.jwp.http.HttpRequest;
-import nextstep.jwp.http.HttpResponse;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
 
 public interface HttpRequestHandler {
