+++ tomcat/src/main/java/nextstep/jwp/handler/RegisterRequestHandler.java
@@ -3,11 +3,11 @@ package nextstep.jwp.handler;
 import java.util.UUID;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.http.HttpCookie;
-import nextstep.jwp.http.HttpRequest;
-import nextstep.jwp.http.HttpRequestBody;
-import nextstep.jwp.http.HttpResponse;
 import nextstep.jwp.http.HttpVersion;
 import nextstep.jwp.http.Location;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.request.HttpRequestBody;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.model.User;
 import nextstep.jwp.util.ResourcesUtil;
