+++ tomcat/src/main/java/nextstep/jwp/http/request/HttpRequestBody.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.http.request;
 
 import java.util.Map;
 import java.util.Objects;
