+++ tomcat/src/main/java/nextstep/jwp/http/session/SessionManager.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.http.session;
 
 import java.util.Map;
 import java.util.Optional;
