+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -13,13 +13,13 @@ import nextstep.jwp.handler.LoginRequestHandler;
 import nextstep.jwp.handler.RegisterRequestHandler;
 import nextstep.jwp.http.ContentType;
 import nextstep.jwp.http.HttpCookie;
-import nextstep.jwp.http.HttpRequest;
-import nextstep.jwp.http.HttpRequestBody;
-import nextstep.jwp.http.HttpRequestHeaders;
-import nextstep.jwp.http.HttpResponse;
 import nextstep.jwp.http.HttpStatus;
 import nextstep.jwp.http.HttpVersion;
 import nextstep.jwp.http.Location;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.request.HttpRequestBody;
+import nextstep.jwp.http.request.HttpRequestHeaders;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.util.ResourcesUtil;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
