+++ tomcat/src/test/java/nextstep/jwp/http/request/HttpRequestTest.java
@@ -1,10 +1,13 @@
-package nextstep.jwp.http;
+package nextstep.jwp.http.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import nextstep.jwp.exception.UncheckedServletException;
+import nextstep.jwp.http.ContentType;
+import nextstep.jwp.http.HttpCookie;
+import nextstep.jwp.http.HttpMethod;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
