+++ tomcat/src/test/java/nextstep/jwp/http/request/RequestParamsTest.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.http.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
